refactor: useCardState 디렉토리 변경
