+++ src/pages/AddCard/hooks/useCardExpiration.ts
@@ -1,5 +1,5 @@
 import { useCallback } from 'react';
-import { useCardState } from '../../../hooks/useCardState';
+import { useCardState } from '../../../providers/CardState/hooks/useCardState';
 
 const EXPIRATION_MAX_LENGTH = 2;
 const MAX_MONTH = 12;
