+++ src/pages/AddCard/hooks/useCardOwner.ts
@@ -1,5 +1,5 @@
 import { useCallback } from 'react';
-import { useCardState } from '../../../hooks/useCardState';
+import { useCardState } from '../../../providers/CardState/hooks/useCardState';
 
 const useCardOwner = () => {
   const { cardState, setCardState } = useCardState();
