+++ src/pages/AddCard/hooks/useCardPassword.ts
@@ -1,5 +1,5 @@
 import { useCallback } from 'react';
-import { useCardState } from '../../../hooks/useCardState';
+import { useCardState } from '../../../providers/CardState/hooks/useCardState';
 
 const PASSWORD_INPUT_MAX_LENGTH = 2;
 const useCardPassword = () => {
