+++ src/pages/AddCard/hooks/useCardSecurityCode.ts
@@ -1,5 +1,5 @@
 import { useCallback } from 'react';
-import { useCardState } from '../../../hooks/useCardState';
+import { useCardState } from '../../../providers/CardState/hooks/useCardState';
 
 const SECURITY_CODE_MAX_LENGTH = 3;
