+++ src/pages/AddCardSuccess/hooks/useNickname.ts
@@ -1,5 +1,5 @@
 import { useCallback } from 'react';
-import { useCardState } from '../../../hooks/useCardState';
+import { useCardState } from '../../../providers/CardState/hooks/useCardState';
 
 const useCardNickname = () => {
   const { cardState, setCardState } = useCardState();
